TERA-432: Implement ignoreList logic & refactoring
